requirements.txt
pyserial
unittest
platformio